GB - Алгоритмы Урок 2 Задание 1 - в разработке.
