+++ JobLessonAlgo02v01Part01/JobLessonAlgo02v01Part01.csproj
@@ -43,6 +43,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ILinkedList.cs" />
+    <Compile Include="Node.cs" />
+    <Compile Include="NodesFunctions.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
